+++ packages/react/src/components/Table/Table.jsx
@@ -16,7 +16,6 @@ import deprecate from '../../internal/deprecate';
 
 import {
   TableColumnsPropTypes,
-  TableRowPropTypes,
   ExpandedRowsPropTypes,
   EmptyStatePropTypes,
   TableSearchPropTypes,
@@ -28,6 +27,7 @@ import {
   TableOrderingPropType,
   TableFiltersPropType,
   TableToolbarActionsPropType,
+  TableRowsPropTypes,
 } from './TablePropTypes';
 import TableHead from './TableHead/TableHead';
 import TableToolbar from './TableToolbar/TableToolbar';
@@ -60,7 +60,7 @@ const propTypes = {
   /** Specify the properties of each column group in the table. Defaults to empty array. */
   columnGroups: TableColumnGroupPropType,
   /** Row value data for the body of the table */
-  data: TableRowPropTypes.isRequired,
+  data: TableRowsPropTypes.isRequired,
   /** Expanded data for the table details */
   expandedData: ExpandedRowsPropTypes,
 
@@ -745,7 +745,7 @@ const Table = (props) => {
   const totalColumns =
     visibleColumns.length +
     (hasMultiSelect ? 1 : 0) +
-    (options.hasRowExpansion ? 1 : 0) +
+    (options.hasRowExpansion || options.hasRowNesting ? 1 : 0) +
     (options.hasRowActions ? 1 : 0) +
     (showExpanderColumn ? 1 : 0);
 
@@ -926,6 +926,7 @@ const Table = (props) => {
                     }}
                     // TODO: remove id in V3.
                     data-testid={`${id || testId}-filter-tag-${advancedFilter.filterId}`}
+                    type="blue"
                   >
                     {advancedFilter.filterTitleText}
                   </Tag>